Unify the two copies of cleanKind
So the code coverage does not drop too much.

At the same time, it's clear the heuristic used is pretty gross,
and we should use the #_[canonical] annotation on the class instead.
