+++ lib/ast/class_def.js
@@ -14,7 +14,7 @@ const assert = require('assert');
 //const Ast = require('./ast');
 const Type = require('../type');
 const { prettyprintClassDef } = require('../prettyprint');
-const { clean } = require('../utils');
+const { clean, cleanKind } = require('../utils');
 const { Value } = require('./values');
 const { Statement, InputParam } = require('./program');
 const { FunctionDef } = require('./function_def');
@@ -22,28 +22,6 @@ const { getString, extractImports, typeToHTML } = require('./manifest_utils');
 const toJS = require('./toJS');
 
 // Class definitions
-
-function cleanKind(kind) {
-    // thingengine.phone -> phone
-    if (kind.startsWith('org.thingpedia.builtin.thingengine.'))
-        kind = kind.substr('org.thingpedia.builtin.thingengine.'.length);
-    // org.thingpedia.builtin.omlet -> omlet
-    if (kind.startsWith('org.thingpedia.builtin.'))
-        kind = kind.substr('org.thingpedia.builtin.'.length);
-    // org.thingpedia.weather -> weather
-    if (kind.startsWith('org.thingpedia.'))
-        kind = kind.substr('org.thingpedia.'.length);
-    // com.xkcd -> xkcd
-    if (kind.startsWith('com.'))
-        kind = kind.substr('com.'.length);
-    if (kind.startsWith('gov.'))
-        kind = kind.substr('gov.'.length);
-    if (kind.startsWith('org.'))
-        kind = kind.substr('org.'.length);
-    if (kind.startsWith('uk.co.'))
-        kind = kind.substr('uk.co.'.length);
-    return clean(kind);
-}
 
 /**
  * The definition of a ThingTalk class.
