+++ lib/describe.js
@@ -13,7 +13,7 @@ const assert = require('assert');
 
 const Ast = require('./ast');
 const Type = require('./type');
-const { clean } = require('./utils');
+const { clean, cleanKind } = require('./utils');
 const FormatUtils = require('./format_utils');
 const { Currency } = require('./builtin/values');
 const I18n = require('./i18n');
@@ -856,9 +856,9 @@ class Describer {
 
             switch (functionType) {
             case 'query':
-                return this._("your %s").format(doCapitalizeSelector(kind));
+                return this._("your %s").format(capitalize(cleanKind(kind)));
             case 'action':
-                return this._("perform any action on your %s").format(doCapitalizeSelector(kind));
+                return this._("perform any action on your %s").format(capitalize(cleanKind(kind)));
             default:
                 return '';
             }
@@ -981,7 +981,7 @@ class Describer {
 }
 
 function capitalize(str) {
-    return (str[0].toUpperCase() + str.substr(1)).replace(/[.\-_]([a-z])/g, (whole, char) => ' ' + char.toUpperCase()).replace(/[.\-_]/g, '');
+    return str.split(/\s+/g).map((word) => word[0].toUpperCase() + word.substring(1)).join(' ');
 }
 
 function capitalizeSelector(prim) {
@@ -994,27 +994,10 @@ function capitalizeSelector(prim) {
 }
 
 function doCapitalizeSelector(kind, channel) {
-    // thingengine.phone -> phone
-    if (kind.startsWith('org.thingpedia.builtin.thingengine.'))
-        kind = kind.substr('org.thingpedia.builtin.thingengine.'.length);
-    // org.thingpedia.builtin.omlet -> omlet
-    if (kind.startsWith('org.thingpedia.builtin.'))
-        kind = kind.substr('org.thingpedia.builtin.'.length);
-    // org.thingpedia.weather -> weather
-    if (kind.startsWith('org.thingpedia.'))
-        kind = kind.substr('org.thingpedia.'.length);
-    // com.xkcd -> xkcd
-    if (kind.startsWith('com.'))
-        kind = kind.substr('com.'.length);
-    if (kind.startsWith('gov.'))
-        kind = kind.substr('gov.'.length);
-    if (kind.startsWith('org.'))
-        kind = kind.substr('org.'.length);
-    if (kind.startsWith('uk.co.'))
-        kind = kind.substr('uk.co.'.length);
+    kind = cleanKind(kind);
 
     if (kind === 'builtin' || kind === 'remote' || kind.startsWith('__dyn_'))
-        return capitalize(channel);
+        return capitalize(clean(channel));
     else
         return capitalize(kind);
 }
