+++ lib/utils.js
@@ -23,6 +23,29 @@ function clean(name) {
     return name.replace(/_/g, ' ').replace(/([^A-Z ])([A-Z])/g, '$1 $2').toLowerCase();
 }
 
+function cleanKind(kind) {
+    // thingengine.phone -> phone
+    if (kind.startsWith('org.thingpedia.builtin.thingengine.'))
+        kind = kind.substr('org.thingpedia.builtin.thingengine.'.length);
+    // org.thingpedia.builtin.omlet -> omlet
+    if (kind.startsWith('org.thingpedia.builtin.'))
+        kind = kind.substr('org.thingpedia.builtin.'.length);
+    // org.thingpedia.weather -> weather
+    if (kind.startsWith('org.thingpedia.'))
+        kind = kind.substr('org.thingpedia.'.length);
+    // com.xkcd -> xkcd
+    if (kind.startsWith('com.'))
+        kind = kind.substr('com.'.length);
+    if (kind.startsWith('gov.'))
+        kind = kind.substr('gov.'.length);
+    if (kind.startsWith('org.'))
+        kind = kind.substr('org.'.length);
+    if (kind.startsWith('uk.co.'))
+        kind = kind.substr('uk.co.'.length);
+    kind = kind.replace(/[.-]/g, ' ');
+    return clean(kind);
+}
+
 // this regexp is similar to the one in runtime/formatter.js, but it does not allow '%' as an option
 // FIXME: unify
 const PARAM_REGEX = /\$(?:\$|([a-zA-Z0-9_]+(?![a-zA-Z0-9_]))|{([a-zA-Z0-9_]+)(?::([a-zA-Z0-9_]+))?})/;
@@ -52,6 +75,7 @@ module.exports = {
     split,
     makeIndex,
     clean,
+    cleanKind,
 
     getSchemaForSelector(schemaRetriever, type, name, schemaType, getMeta = false, classes = {}) {
         if (type in classes) {
